pyproject.toml
[project]
name = "sku-llm-match"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cozeloop>=0.1.19",
    "jieba>=0.42.1",
    "jupyter>=1.1.1",
    "langchain-openai>=1.0.2",
    "langgraph>=1.0.3",
    "openpyxl>=3.1.5",
    "opik>=1.9.14",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
]
